refactor serviceDetailRow out from serviceDetail
