+++ app/components/service/ServiceDetail.js
@@ -2,9 +2,9 @@ import React from 'react'
 import Parse from 'parse'
 import { Link } from 'react-router'
 
-import Colors from '../../constants/JendelaColors'
 import Loading from '../template/Loading'
 import ServiceSummary from './ServiceSummary'
+import ServiceDetailRow from './ServiceDetailRow'
 
 const styles = {
     content: {
@@ -13,37 +13,10 @@ const styles = {
     }
 }
 
-class ServiceDetailRow extends React.Component {
-    render() {
-        const styles = {
-            container: {
-                marginBottom: "1.2em"
-            },
-            title: {
-                fontWeight: 900,
-                textTransform: "uppercase",
-                color: "#88bcb4",
-            },
-        }
-        return (
-            <div className="row" style={styles.container}>
-                <div className="small-12 medium-4 columns">
-                    <h6 style={styles.title}>{this.props.title}</h6>
-                </div>
-                <div style={{ paddingBottom: "10px" }} className="small-12 medium-8 columns">
-                    {this.props.children}
-                </div>
-            </div>
-        )
-    }
-}
-ServiceDetailRow.defaultProps = { title: '' }
-
 class ServiceDetail extends React.Component {
 
     constructor(props) {
-        super(props);
-
+        super(props)
         this.state = { "services": [] }
         this.componentWillMount.bind(this);
     }
@@ -51,7 +24,7 @@ class ServiceDetail extends React.Component {
     componentWillMount() {
         new Parse.Query('Service').get(this.props.params.serviceId).then((service) => {
             this.setState({"service": service});
-        });
+        })
     }
 
     render() {
@@ -95,8 +68,7 @@ class ServiceDetail extends React.Component {
 }
 
 ServiceDetail.defaultProps = {
-    params: {
-        "serviceId": undefined
-    }
-};
+    params: { "serviceId": undefined }
+}
+
 export default ServiceDetail
